ai-service/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
hypothesis
pytest
httpx
python-dotenv
